YARN-679 cleanup of service launcher tests
