+++ hadoop-common-project/hadoop-common/src/test/java/org/apache/hadoop/service/launcher/TestServiceLauncherCreationFailures.java
@@ -27,7 +27,6 @@ import org.junit.Test;
 public class TestServiceLauncherCreationFailures extends
     AbstractServiceLauncherTestBase {
 
-
   public static final String SELF =
       "org.apache.hadoop.service.launcher.TestServiceLauncherCreationFailures";
